Move ndcToCameraRay
